add and remove functions  done in a 100%
