+++ src/index.js
@@ -1,5 +1,8 @@
 import './styles/index.css';
-import Task from './Task';
+import addtask from './addtaskfunction.js';
+import editTask from './editfunction.js';
+import deleteTask from './delatefunction.js';
+import getData from './getdatafunction.js';
 
 const listContainer = document.querySelector('.list-container');
 const inputField = document.querySelector('input');
@@ -7,53 +10,7 @@ const tasks = document.querySelector('.tasks');
 const button = document.querySelector('button');
 const data = JSON.parse(localStorage.getItem('list'));
 
-// add task to the UI functionality
-const listArray = [];
-
-const addTask = (taskValue) => {
-  const listItem = document.createElement('li');
-  listItem.innerHTML = `
-  <input type="checkbox" class="checkbox"/>
-  <span>${taskValue}</span >
-  <i class="fa-solid fa-ellipsis-vertical edit-btn"></i>
-  <i class="fa-solid fa-trash-can delete-btn"></i>
-  `;
-  tasks.appendChild(listItem);
-
-  // checkbox functionality
-  const checkbox = document.querySelectorAll('.checkbox');
-  checkbox.forEach((i) => {
-    i.addEventListener('click', () => {
-      i.parentElement.classList.toggle('checkContainer');
-      i.nextElementSibling.classList.toggle('checkTask');
-      i.parentElement.lastElementChild.classList.toggle('active');
-      i.parentElement.lastElementChild.previousElementSibling.classList.toggle('disable');
-      updateStorage();
-    });
-  });
-
-  // Add data to local storage
-  const storedData = new Task(taskValue, false, checkbox.length - 1);
-  listArray.push(storedData);
-  localStorage.setItem('list', JSON.stringify(listArray));
-
-  // Edit event listener
-  const editBtn = document.querySelectorAll('.edit-btn');
-  editBtn.forEach((i) => {
-    i.addEventListener('click', () => {
-      editTask(listItem, i.previousElementSibling);
-      // updateStorage();
-    });
-  });
-
-  //  Delete event listener
-  const removeBtn = document.querySelectorAll('.delete-btn');
-  removeBtn.forEach((i) => {
-    i.addEventListener('click', () => {
-      deleteTask(i.parentElement);
-    });
-  });
-};
+addtask();
 
 // add task event listener
 inputField.addEventListener('keypress', (e) => {
@@ -64,84 +21,13 @@ inputField.addEventListener('keypress', (e) => {
   }
 });
 
-// edit functionality
-const editTask = (listItem, task) => {
-  const editInput = document.createElement('input');
-  const storedData = JSON.parse(localStorage.getItem('list'));
-  editInput.type = 'text';
-  editInput.className = 'editInput';
-  editInput.value = task.textContent;
-  listItem.replaceChild(editInput, task);
-  editInput.addEventListener('keypress', (e) => {
-    if (e.key === 'Enter' && editInput.value) {
-      const result = storedData.filter((text) => text.description === task.textContent);
-      const emptyArray = [];
-      for (let i = 0; i < storedData.length; i += 1) {
-        if (storedData[i].index === result[0].index) {
-          storedData[i].description = editInput.value;
-        }
-        emptyArray.push(storedData[i]);
-        localStorage.setItem('list', JSON.stringify(emptyArray));
-      }
-      listItem.replaceChild(task, editInput);
-      task.textContent = editInput.value;
-    }
-  });
-};
+editTask();
 
-// Delete task from list
-const deleteTask = (task) => {
-  tasks.removeChild(task);
-  let count = -1;
-  const storedData = JSON.parse(localStorage.getItem('list'));
-  const data = Array.from(storedData).filter((i) => i.completed === false);
-  data.map((i) => i.index = count += 1);
-  localStorage.setItem('list', JSON.stringify(data));
-};
+deleteTask();
 
-// Get data from local storage
-const getData = () => {
-  data.map((i) => {
-    listArray.push(i);
-    const listItem = document.createElement('li');
-    listItem.innerHTML = `
-  <input type="checkbox" class="checkbox"/>
-  <span>${i.description}</span >
-  <i class="fa-solid fa-ellipsis-vertical edit-btn"></i>
-  <i class="fa-solid fa-trash-can delete-btn"></i>
-  `;
-    tasks.appendChild(listItem);
+getData();
 
-    const editBtn = document.querySelectorAll('.edit-btn');
-    editBtn.forEach((i) => {
-      i.addEventListener('click', () => {
-        editTask(listItem, i.previousElementSibling);
-      });
-    });
-  });
-
-  const checkbox = document.querySelectorAll('.checkbox');
-  checkbox.forEach((i) => {
-    i.addEventListener('click', () => {
-      i.parentElement.classList.toggle('checkContainer');
-      i.nextElementSibling.classList.toggle('checkTask');
-      i.parentElement.lastElementChild.classList.toggle('active');
-      i.parentElement.lastElementChild.previousElementSibling.classList.toggle('disable');
-      updateStorage();
-    });
-  });
-
-  const removeBtn = document.querySelectorAll('.delete-btn');
-  removeBtn.forEach((i) => {
-    i.addEventListener('click', () => {
-      deleteTask(i.parentElement);
-    });
-  });
-
-  localStorage.setItem('list', JSON.stringify(listArray));
-};
-
-window.addEventListener('load', getData);
+window.addEventListener('load', getData());
 
 const updateStorage = () => {
   const localData = JSON.parse(localStorage.getItem('list'));
@@ -172,3 +58,4 @@ const clearAll = () => {
 };
 
 button.addEventListener('click', clearAll);
+export { updateStorage, listContainer, data, tasks };
